% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_airbnb_data.R
\name{clean_airbnb_data}
\alias{clean_airbnb_data}
\title{Clean raw Airbnb listing data for EDA and modelling in script 02_clean_airbnb_data.R}
\usage{
clean_airbnb_data(raw_airbnb)
}
\arguments{
\item{raw_airbnb}{A data frame or data frame extension (e.g. a tibble)}
}
\value{
A data frame object containing cleaned data of airbnb listings:
- selected specific set of features for future EDA
- filtered out entries (rows) with NA values
- refactored certain char type features to categories
}
\description{
Return a clean dataframe
}
\examples{
raw_airbnb <- data.frame(id = 1, host_is_superhost = TRUE, 
                         property_type = "House", neighbourhood_cleansed = "Downtown",
                         price = "$150", review_scores_rating = 4.5, 
                         reviews_per_month = 3, accommodates = 2, bedrooms = 1,
                         bathrooms = 1, room_type = "Entire")
                         
clean_airbnb_data(raw_airbnb)
}
