% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_columns_to_factor.R
\name{convert_columns_to_factor}
\alias{convert_columns_to_factor}
\title{Convert selected columns to factor type}
\usage{
convert_columns_to_factor(df)
}
\arguments{
\item{df}{A data frame containing the columns: room_type, host_is_superhost, city, weekdays}
}
\value{
A data frame with the specified columns converted to factors
}
\description{
This function takes a data frame and converts specified columns to factors.
}
