% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_combine_data.R
\name{download_combine_data}
\alias{download_combine_data}
\title{This function downloads Airbnb data from given URLs, adds metadata, and combines them.}
\usage{
download_combine_data(urls)
}
\arguments{
\item{urls}{A named list where names indicate the city and time period (weekdays/weekends)
and values are the corresponding dataset URLs.}
}
\value{
A combined data frame with an added city and weekdays column.
}
\description{
This function downloads Airbnb data from given URLs, adds metadata, and combines them.
}
