% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_train_test.R
\name{split_train_test}
\alias{split_train_test}
\title{Split data into training and testing sets}
\usage{
split_train_test(data, response_col, train_ratio = 0.8, seed = 123)
}
\arguments{
\item{data}{A data frame containing the dataset to split}

\item{response_col}{A string specifying the name of the response column}

\item{train_ratio}{A numeric value indicating the proportion of data to be used for training (default is 0.8)}

\item{seed}{An integer seed for reproducibility (default is 123)}
}
\value{
A list containing two elements: 'train' and 'test', which are data frames
}
\description{
This function splits a dataset into training and testing sets using stratified sampling
based on the response variable.
}
\examples{
df <- data.frame(
  x = rnorm(100),
  y = rnorm(100)
)
split_data <- split_train_test(df, response_col = "y", train_ratio = 0.75)
str(split_data$train)
str(split_data$test)
}
