% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train a regression model using caret}
\usage{
train_model(train_data)
}
\arguments{
\item{train_data}{A data frame containing the training set. Must include a column called \code{price}.}
}
\value{
A trained caret model.
}
\description{
Train a regression model using caret
}
