% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_airbnb_data.R
\name{validate_airbnb_data}
\alias{validate_airbnb_data}
\title{Convert selected columns to factor type}
\usage{
validate_airbnb_data(df)
}
\arguments{
\item{df}{A data frame containing the columns: room_type, host_is_superhost, city, weekdays}
}
\value{
A data frame with the specified columns converted to factors
}
\description{
This function takes a data frame and converts specified columns to factors.
}
